refactor: move parseAmount to global scope
