+++ .github/CODEOWNERS
@@ -49,6 +49,8 @@ metro.transform.js @MetaMask/mobile-platform @MetaMask/core-platform
 # Ramps Team
 app/components/UI/Ramp/              @MetaMask/ramp
 app/reducers/fiatOrders/             @MetaMask/ramp
+app/util/parseAmount.ts              @MetaMask/ramp
+app/util/parseAmount.test.ts         @MetaMask/ramp
 
 # Card Team
 app/components/UI/Card/              @MetaMask/card
@@ -98,7 +100,9 @@ app/components/UI/EvmAccountSelectorList                    @MetaMask/accounts-e
 
 # Swaps Team
 app/components/UI/Swaps              @MetaMask/swaps-engineers
-app/components/UI/Bridge              @MetaMask/swaps-engineers
+app/components/UI/Bridge             @MetaMask/swaps-engineers
+app/util/parseAmount.ts              @MetaMask/swaps-engineers
+app/util/parseAmount.test.ts         @MetaMask/swaps-engineers
 
 # Notifications Team
 app/components/Views/Notifications @MetaMask/notifications
